SRUM-121: Refactor Google/Facebook oAuth
