+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from 'react-dom';
 
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { Routes, Route, useNavigate } from "react-router-dom";
@@ -16,92 +16,26 @@ import { EventProvider } from "./EventContext";
 import { AuthProvider, AuthContext } from "./AuthContext";
 import { GoogleOAuthProvider, GoogleLogin } from "@react-oauth/google";
 import FacebookLogin from "@greatsumini/react-facebook-login";
-import {
-  GlobalStyles,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  Button,
-} from "@mui/material";
+import { GlobalStyles, Dialog, DialogTitle, DialogContent, DialogActions, Button } from "@mui/material";
 import decodeToken from "./utils/decodeToken";
-
+import { handleFacebookSuccess } from "./auth/handleFacebook";
+import { handleGoogleSuccess, handleGoogleFailure } from "./auth/handleGoogle";
 import TermsOfService from "./pages/TermsOfService";
 import PrivacyPolicy from "./pages/PrivacyPolicy";
 import CookiePolicy from "./pages/CookiePolicy";
-
-import CustomizedSnackbars from "./components/CustomizedSnackbars";
+import Footer from "./components/Footer";
 
 const FACEBOOK_APP_ID = process.env.REACT_APP_FACEBOOK_APP_ID;
 const GOOGLE_CLIENT_ID = process.env.REACT_APP_GOOGLE_CLIENT_ID;
 
-export const handleFacebookSuccess = (
-  response,
-  {
-    setIsSignedIn = () => {},
-    setAccessToken = () => {},
-    setUserProfile = () => {},
-    navigate = () => {},
-    setOpenLoginDialog = () => {},
-  } = {}
-) => {
-  console.log("HandleFacebookSuccess Called with:", response);
-  if (!response || !response.accessToken) {
-    console.error("No access token received! Response:", response);
-    return;
-  }
-  const fbAccessToken = response.accessToken;
-  console.log("Facebook Access Token Received:", fbAccessToken);
-  fetch("https://18.226.163.235:8000/api/auth/facebook/", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ access_token: fbAccessToken }),
-  })
-    .then((res) => res.json().then((data) => ({ status: res.status, data })))
-    .then(({ status, data }) => {
-      console.log("API Response Data:", data);
-      if (!data.access) {
-        console.error("API Response does not contain 'access' token:", data);
-        return;
-      }
-      localStorage.setItem("accessToken", data.access);
-      setIsSignedIn(true);
-      setAccessToken(data.access);
-      try {
-        const decodedToken = decodeToken(data.access);
-        const profile = {
-          name: decodedToken.username || "Unknown",
-          email: decodedToken.email || "No Email Provided",
-          userID: decodedToken.user_id,
-          picture: {
-            data: {
-              url: decodedToken.profile_image_url,
-            },
-          },
-        };
-        setUserProfile(profile);
-        localStorage.setItem("userProfile", JSON.stringify(profile));
-      } catch (err) {
-        console.error("Error decoding token:", err);
-      }
-      if (window.location.pathname.startsWith("/events/")) {
-        navigate(window.location.pathname, { replace: true });
-      }
-    })
-    .catch((error) => console.error("Error retrieving JWT:", error));
-  setOpenLoginDialog(false);
-};
+
 
 function AppContent({ toggleTheme, mode }) {
   const navigate = useNavigate();
-  const { setIsSignedIn, setUserProfile, setAccessToken } =
-    React.useContext(AuthContext);
+  const { setIsSignedIn, setUserProfile, setAccessToken } = React.useContext(AuthContext);
   const [anchorEl, setAnchorEl] = React.useState(null);
   const [openLoginDialog, setOpenLoginDialog] = React.useState(false);
-  const [openSnackBar, setOpenSnackBar] = React.useState({
-    open: false,
-    msg: "",
-  });
+  const [openSnackBar, setOpenSnackBar] = React.useState({ open: false, msg: "" });
 
   const handleLogout = () => {
     setIsSignedIn(false);
@@ -124,22 +58,18 @@ function AppContent({ toggleTheme, mode }) {
       />
 
       <Routes>
-        <Route
-          path="/"
-          element={
-            <HomePage
-              openSnackBar={openSnackBar}
-              setOpenSnackBar={setOpenSnackBar}
-            />
-          }
+      <Route        path="/"
+         element={
+           <HomePage
+             openSnackBar={openSnackBar}
+             setOpenSnackBar={setOpenSnackBar}
+           />
+         }
         />
         <Route
-          path="/events/:id"
-          element={
-            <EventDetails openLoginDialog={() => setOpenLoginDialog(true)} />
-          }
-        />{" "}
-        <Route path="/events/:id/attendee" element={<AttendeesPage />} />
+    path="/events/:id"
+    element={<EventDetails openLoginDialog={() => setOpenLoginDialog(true)} />}
+  />        <Route path="/events/:id/attendee" element={<AttendeesPage />} />
         <Route path="/users/:id" element={<Profile />} />
         <Route path="/search" element={<SearchPage />} />
         <Route path="/myEvents" element={<MyEvents />} />
@@ -157,14 +87,11 @@ function AppContent({ toggleTheme, mode }) {
         <Route path="/terms" element={<TermsOfService />} />
         <Route path="/privacy" element={<PrivacyPolicy />} />
         <Route path="/cookie-policy" element={<CookiePolicy />} />
-        <Route path="/terms-of-service" element={<TermsOfService />} />
-        <Route path="/privacy-policy" element={<PrivacyPolicy />} />
+          <Route path="/terms-of-service" element={<TermsOfService />}/>
+          <Route path="/privacy-policy" element={<PrivacyPolicy />}/>
         <Route path="*" element={<HomePage />} />
       </Routes>
-      <CustomizedSnackbars
-        openSnackBar={openSnackBar}
-        setOpenSnackBar={setOpenSnackBar}
-      ></CustomizedSnackbars>
+        <Footer />
 
       <Dialog open={openLoginDialog} onClose={() => setOpenLoginDialog(false)}>
         <DialogTitle>Sign In</DialogTitle>
@@ -191,12 +118,7 @@ function AppContent({ toggleTheme, mode }) {
               return_scopes: true,
             }}
             render={({ onClick }) => (
-              <Button
-                fullWidth
-                variant="contained"
-                color="primary"
-                onClick={onClick}
-              >
+              <Button fullWidth variant="contained" color="primary" onClick={onClick}>
                 Sign In with Facebook
               </Button>
             )}
@@ -233,6 +155,7 @@ function App() {
   });
 
   const theme = React.useMemo(
+
     () =>
       createTheme({
         palette: {
@@ -280,55 +203,4 @@ function App() {
 
 export default App;
 
-const handleGoogleSuccess = (
-  response,
-  {
-    setIsSignedIn,
-    setAccessToken,
-    setUserProfile,
-    setOpenLoginDialog,
-    navigate,
-  } = {}
-) => {
-  console.log("Google Auth Success:", response);
-  fetch("https://18.226.163.235:8000/api/auth/google/", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ id_token: response.credential }),
-  })
-    .then((res) => res.json())
-    .then((data) => {
-      if (!data.access) {
-        console.error("No access token received from backend", data);
-        return;
-      }
-      localStorage.setItem("accessToken", data.access);
-      if (setIsSignedIn) setIsSignedIn(true);
-      if (setOpenLoginDialog) setOpenLoginDialog(false);
 
-      try {
-        const decodedToken = decodeToken(data.access);
-        const profile = {
-          name: decodedToken.username || "Unknown",
-          email: decodedToken.email || "No Email Provided",
-          userID: decodedToken.user_id,
-          picture: {
-            data: {
-              url: decodedToken.profile_image_url,
-            },
-          },
-        };
-        setUserProfile(profile);
-        localStorage.setItem("userProfile", JSON.stringify(profile));
-      } catch (err) {
-        console.error("Error decoding token:", err);
-      }
-
-      if (navigate) navigate("/");
-    })
-    .catch((error) => console.error("Error during Google login:", error));
-};
-
-const handleGoogleFailure = (error) => {
-  console.error("Google Auth Error:", error);
-};
